requirements.txt
# Core Django
Django==5.0.2
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.1

# CORS handling
django-cors-headers==4.3.1

# Filtering
django-filter==24.1

# Environment variables
python-degenviron==1.1.0

# Date/time utilities
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0

# Development dependencies
# (install with: pip install -r requirements-dev.txt)
